+++ example/lib/main.dart
@@ -6,7 +6,7 @@ void main() {
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}): super(key: key);
+  const MyApp({Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -49,44 +49,51 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  Widget previewList(){
+  Widget previewList() {
     return SizedBox(
       height: 96,
       child: ListView(
         scrollDirection: Axis.horizontal,
         shrinkWrap: true,
-        children: List.generate(mediaList.length, (index) => Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: SizedBox(
-            height: 80,
-            width: 80,
-            child: Image.memory(mediaList[index].thumbnail, fit: BoxFit.cover,),
-          ),
-        )),
+        children: List.generate(
+            mediaList.length,
+            (index) => Padding(
+                  padding: const EdgeInsets.all(8.0),
+                  child: SizedBox(
+                    height: 80,
+                    width: 80,
+                    child: Image.memory(
+                      mediaList[index].thumbnail,
+                      fit: BoxFit.cover,
+                    ),
+                  ),
+                )),
       ),
     );
   }
 
-  void openImagePicker(BuildContext context){
+  void openImagePicker(BuildContext context) {
     // openCamera(onCapture: (image){
     //   setState(()=> mediaList = [image]);
     // });
-    showModalBottomSheet(context: context, builder: (context){
-      return MediaPicker(
-        mediaList: mediaList,
-        onPick: (selectedList){
-          setState(()=> mediaList = selectedList);
-          Navigator.pop(context);
-        },
-        onCancel: ()=> Navigator.pop(context),
-        mediaCount: MediaCount.multiple,
-        mediaType: MediaType.image,
-        decoration: PickerDecoration(
-          actionBarPosition: ActionBarPosition.top,
-          blurStrength: 2,
-          completeText: 'Next',
-        ),
-      );
-    });
+    showModalBottomSheet(
+        context: context,
+        builder: (context) {
+          return MediaPicker(
+            mediaList: mediaList,
+            onPick: (selectedList) {
+              setState(() => mediaList = selectedList);
+              Navigator.pop(context);
+            },
+            onCancel: () => Navigator.pop(context),
+            mediaCount: MediaCount.multiple,
+            mediaType: MediaType.image,
+            decoration: PickerDecoration(
+              actionBarPosition: ActionBarPosition.top,
+              blurStrength: 2,
+              completeText: 'Next',
+            ),
+          );
+        });
   }
 }