+++ lib/src/media_list.dart
@@ -39,7 +39,8 @@ class _MediaListState extends State<MediaList> {
     album = widget.album;
     if (widget.mediaCount == MediaCount.multiple) {
       selectedMedias.addAll(widget.previousList);
-      WidgetsBinding.instance!.addPostFrameCallback((_) => widget.headerController.updateSelection!(selectedMedias));
+      WidgetsBinding.instance!.addPostFrameCallback(
+          (_) => widget.headerController.updateSelection!(selectedMedias));
     }
     _fetchNewMedia();
     super.initState();
@@ -56,7 +57,8 @@ class _MediaListState extends State<MediaList> {
       child: GridView.builder(
         controller: widget.scrollController,
         itemCount: _mediaList.length,
-        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: widget.decoration!.columnCount),
+        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
+            crossAxisCount: widget.decoration!.columnCount),
         itemBuilder: (BuildContext context, int index) {
           return _mediaList[index];
         },
@@ -97,7 +99,8 @@ class _MediaListState extends State<MediaList> {
             if (isSelected)
               setState(() => selectedMedias.add(media));
             else
-              setState(() => selectedMedias.removeWhere((_media) => _media.id == media.id));
+              setState(() => selectedMedias
+                  .removeWhere((_media) => _media.id == media.id));
             widget.headerController.updateSelection!(selectedMedias);
           },
           isSelected: isPreviouslySelected(asset),